+++ src/modules/core/GameConfig.h
@@ -102,5 +102,8 @@ constexpr const char *VoxformatQBSaveLeftHanded = "voxformat_qbsavelefthanded";
 constexpr const char *VoxformatQBSaveCompressed = "voxformat_qbsavecompressed";
 constexpr const char *VoxFormatGLTF_KHR_materials_pbrSpecularGlossiness = "voxformat_gltf_khr_materials_pbrspecularglossiness";
 constexpr const char *VoxFormatGLTF_KHR_materials_specular = "voxformat_gltf_khr_materials_specular";
+constexpr const char *VoxformatImageVolumeMaxDepth = "voxformat_imagevolumemaxdepth";
+constexpr const char *VoxformatImageVolumeBothSides = "voxformat_imagevolumebothsides";
+constexpr const char *VoxformatImageImportType = "voxformat_imageimporttype";
 
 }